+++ R/get_boundary_polygons.R
@@ -11,14 +11,35 @@
 #' \cr
 #' Could be
 #' \code{get(data("boundaries", package = "treePlotArea"))}.
+#' @param stop_on_error Throw an error if invalid boundaries are encountered?
+#' (There is at least one boundary in the 2022 survey,
+#' tract 2607, corner 2, boundary 1, which exactly
+#' splits the plot. There is no stand defined that way!).
+#' @param clean_data Omit invalid boundaries in any further calculations?
 #' @return A list with all boundary polygons for each corner for each tract.
 #' @export
 #' @examples
 #' boundaries <- get(data("boundaries", package = "treePlotArea"))
-#' boundary_polygons <- get_boundary_polygons(boundaries)
-get_boundary_polygons <- function(boundaries) {
+#' # For the boundary in tract 2607, corner 2 the azimuth values in `spa_gon`
+#' # and `spk_gon` are identical:
+#' subset(boundaries, tnr == 2607) # It runs exactly through the plot!
+#' try(boundary_polygons <- get_boundary_polygons(boundaries))
+#' boundary_polygons <- get_boundary_polygons(boundaries, stop_on_error = FALSE)
+#' # There are boundaries in  tract 2607, corner 1, and in corner 2:
+#' length(boundary_polygons[["2607"]][["1"]])
+#' length(boundary_polygons[["2607"]][["2"]])
+#' boundary_polygons <- get_boundary_polygons(boundaries, stop_on_error = FALSE,
+#'                                            clean_data = TRUE)
+#' # Now there's boundaries in in tract 2607, corner 1, but not in corner 2:
+#' length(boundary_polygons[["2607"]][["1"]])
+#' length(boundary_polygons[["2607"]][["2"]])
+get_boundary_polygons <- function(boundaries, stop_on_error = TRUE,
+                                  clean_data = FALSE) {
     options <- get_options("boundaries")
     current_boundaries <- get_current_boundaries(boundaries)
+    current_boundaries <- check_boundaries(current_boundaries,
+                                           stop_on_error = stop_on_error,
+                                           clean_data = clean_data)
     tract <- options[["tract_id"]]
     corner <- options[["corner_id"]]
     res <- fritools::tapply(current_boundaries,