+++ R/secant_intersections.R
@@ -4,6 +4,7 @@
 #' slope-intercept from.
 #' @param a The secant's intercept.
 #' @param b The secant's slope.
+#' @param verbose Be verbose?
 #' @param r The circle's radius
 #' @return A matrix of x und y values. For a tangent, both rows are identical,
 #' for a straight line missing the circle, a matrix of \code{\link{NA}}.
@@ -12,7 +13,7 @@
 #' @examples
 #' secant_intersections(a = 0, b = 1, r = 2)
 #' # A tangent
-#' secant_intersections(a = 2, b = 0, r = 2)
+#' secant_intersections(a = 2, b = 0, r = 2, verbose = TRUE)
 #' # Missing the circle
 #' secant_intersections(a = 3, b = 0, r = 2)
 #' # Creating a circle boundary approximation
@@ -22,7 +23,7 @@
 #' for (i in seq(-1, 1, by = 0.01)) {
 #'          points(secant_intersections(Inf, i, 1), pch = "+")
 #' }
-secant_intersections <- function(a, b, r) {
+secant_intersections <- function(a, b, r, verbose = FALSE) {
     if (is.infinite(a)) {
         x1 <- x2 <- as.numeric(b)
         y1 <- sqrt(abs(r^2 - b^2))
@@ -42,7 +43,7 @@ secant_intersections <- function(a, b, r) {
         if (all(is.na(res))) {
             warning("This line misses the circle!")
         } else {
-            warning("This is a tangent only!")
+            if (isTRUE(verbose)) message("This is a tangent only!")
         }
     }
     return(res)